+++ src/lightly_train/_task_models/dinov3_ltdetr_object_detection/train_model.py
@@ -61,6 +61,7 @@ from lightly_train._task_models.train_model import (
     TrainModel,
     TrainModelArgs,
 )
+from lightly_train._torch_compile import TorchCompileArgs
 from lightly_train.types import ObjectDetectionBatch, PathLike
 
 
@@ -130,6 +131,7 @@ class DINOv3LTDETRObjectDetectionTrain(TrainModel):
     task_model_cls = DINOv3LTDETRObjectDetection
     train_transform_cls = DINOv3LTDETRObjectDetectionTrainTransform
     val_transform_cls = DINOv3LTDETRObjectDetectionValTransform
+    torch_compile_args_cls = TorchCompileArgs
 
     def __init__(
         self,